move with dialog funcs to dialog ts
